enhance: Add proper aborting of runs
Aborting a run is different from "closing" it. Closing a run will result
in an error. Aborting a run will cause it to stop at the the next
available event and not return any error. Instead, the run will have its
text appended with "ABORTED BY USER" and all the chat state will be
preserved.

Signed-off-by: Donnie Adams <donnie@acorn.io>